Expose extended screen handler type creation in SH API
